properly structure project/classes to match rest of solution & also make some public things internal because why not
