+++ Headless.Targeting.CSharp/net8/Scripting/MethodParameter.cs
@@ -1,4 +1,4 @@
 ﻿// ReSharper disable once CheckNamespace
 namespace Headless.Targeting.CSharp.Scripting;
 
-public record MethodParameter(int Order, string? Name, Type? Type);
+internal record MethodParameter(int Order, string? Name, Type? Type);
