+++ Headless.Targeting.CSharp/net8/Scripting/ScriptImplementationDescriptor.cs
@@ -1,4 +1,4 @@
 ﻿// ReSharper disable once CheckNamespace
 namespace Headless.Targeting.CSharp.Scripting;
 
-public record ScriptImplementationDescriptor(MethodBodyImplementation EntryPointBodyType, MethodParameter[] ParameterDescriptors);
+internal record ScriptImplementationDescriptor(MethodBodyImplementation EntryPointBodyType, MethodParameter[] ParameterDescriptors);
