+++ Headless.Targeting.JavaScript/InvocationResult.cs
@@ -1,0 +1,8 @@
+﻿namespace Headless.Targeting.JavaScript;
+
+internal class InvocationResult<TResult>(bool isSuccess, string messages, TResult? result) : IInvocationResult<TResult>
+{
+    public bool IsSuccess { get; } = isSuccess;
+    public string Messages { get; } = messages;
+    public TResult? Result { get; } = result;
+}
